fix: Enable clang-tidy for debug/CI builds and address all warnings
Adds clang-tidy integration to CMakeLists.txt for Debug builds and CI,
with a fatal error if clang-tidy is missing in CI. Excludes checks that
are too noisy for low-level math code (pro-bounds-constant-array-index,
readability-magic-numbers).

Library fixes:
- node.cpp: call Node::ReloadEntityInternal() explicitly in constructor
  to avoid virtual dispatch bypass (clang-analyzer-optin.cplusplus.VirtualCall)
- color.h: fix narrowing conversions in FromARGB/FromRGBA; add NOLINT
  guards for intentional magic numbers (bit-shift byte positions)
- vector.h: add NOLINT guards and static_assert pragma-pack workaround
- vector_serialization.h: replace C-array with std::array, use .at()
- node_image.h: replace C-arrays with std::array for slice storage
- animation_track.cpp: add static_cast<float> on int keyframe comparisons
- animation_track.h: name unnamed parameter in operator= declaration
- node_factory.cpp: static_cast -> dynamic_cast for Group downcast
- group.cpp: static_cast<float> for ticks conversion, fix iterator arithmetic
- font_factory.cpp: const-qualify structured binding
- node.cpp: style fixes (do/while -> while, else-after-return)
- rect.h: simplify IsInRect boolean return

Public API additions:
- NodeRect::IsFilled()
- NodeText::GetHorizontalAlignment(), GetVerticalAlignment(),
  IsDropShadow(), GetDropShadowOffset()
- NodeImage::GetImageScale()

Tests:
- Add test_node_entity.cpp covering entity property initialisation for
  NodeRect, NodeText, and NodeImage on construction

+++ src/nodes/group.cpp
@@ -37,7 +37,8 @@ namespace moth_ui {
 
     void Group::Update(uint32_t ticks) {
         Node::Update(ticks);
-        m_animationClipController->Update(ticks / 1000.0f);
+        //NOLINTNEXTLINE(cppcoreguidelines-avoid-magic-numbers)
+        m_animationClipController->Update(static_cast<float>(ticks) / 1000.0f);
         for (auto&& child : m_children) {
             child->Update(ticks);
         }
@@ -51,7 +52,7 @@ namespace moth_ui {
 
     void Group::AddChild(std::shared_ptr<Node> child, size_t index) {
         if (index != static_cast<size_t>(-1)) {
-            auto it = std::begin(m_children) + index;
+            auto it = std::begin(m_children) + static_cast<int32_t>(index);
             m_children.insert(it, child);
         } else {
             m_children.push_back(child);
